# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для English Bot - системы изучения английского языка с использованием SRS (Spaced Repetition System)",
        "title": "English Bot API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8184",
    "basePath": "/",
    "paths": {
...
        "/app/vocab": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех слов пользователя с статистикой (общее количество карточек, количество готовых к повторению, дата последнего повторения)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocab"
                ],
                "summary": "Получить список слов",
                "responses": {
                    "200": {
                        "description": "Список слов с статистикой",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "405": {
                        "description": "Метод не разрешен",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/app/vocab/{word}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Подтверждение удаления (GET) или удаление слова (POST) из словаря пользователя. Путь: /app/vocab/{word}/confirm_delete или /app/vocab/{word}/delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vocab"
                ],
                "summary": "Удалить слово из словаря",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Английское слово для удаления",
                        "name": "word",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Действие: confirm_delete или delete",
                        "name": "action",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о слове или результат удаления",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Слово не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer JWT token authentication. В Swagger UI введите только токен (без \"Bearer \"), префикс будет добавлен автоматически.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}